fixed generate sitemap
