--- scripts/generate-sitemap.js
@@ -1,20 +1,0 @@
-const fs = require('fs');
-const path = require('path');
-const { sitemap } = require('../src/app/sitemap.ts');
-
-const sitemapContent = sitemap().map(entry => `
-  <url>
-    <loc>${entry.url}</loc>
-    <lastmod>${entry.lastModified.toISOString()}</lastmod>
-    <changefreq>${entry.changeFrequency}</changefreq>
-    <priority>${entry.priority}</priority>
-  </url>
-`).join('');
-
-const sitemapXml = `<?xml version="1.0" encoding="UTF-8"?>
-<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-  ${sitemapContent}
-</urlset>`;
-
-fs.writeFileSync(path.join(__dirname, '../public/sitemap.xml'), sitemapXml, 'utf8');
-console.log('Sitemap generated successfully.');
