+++ scripts/generate-sitemap.mjs
@@ -1,6 +1,11 @@
-import { MetadataRoute } from 'next'
+import fs from 'fs';
+import path from 'path';
+import { fileURLToPath } from 'url';
 
-export default function sitemap(): MetadataRoute.Sitemap {
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+function sitemap() {
   return [
     {
       url: 'https://virtualcoffee.io',
@@ -40,3 +45,23 @@ export default function sitemap(): MetadataRoute.Sitemap {
     },
   ]
 }
+
+console.log('Generating sitemap...');
+
+const sitemapContent = sitemap().map(entry => `
+  <url>
+    <loc>${entry.url}</loc>
+    <lastmod>${entry.lastModified.toISOString()}</lastmod>
+    <changefreq>${entry.changeFrequency}</changefreq>
+    <priority>${entry.priority}</priority>
+  </url>
+`).join('');
+
+const sitemapXml = `<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  ${sitemapContent}
+</urlset>`;
+
+const outputPath = path.join(__dirname, '../public/sitemap.xml');
+fs.writeFileSync(outputPath, sitemapXml, 'utf8');
+console.log(`Sitemap generated successfully at ${outputPath}`);
